Now user can change email,password and delete account from settings

+++ backend/src/controllers/userControllers.js
@@ -32,16 +32,27 @@ export const getUser = async (req, res) => {
 };
 
 export const updateProfile = async (req, res) => {
-  const { oldPassword, newPassword, confirmNewPassword, username, email } = req.body;
+  const { currentPassword, newPassword, confirmNewPassword, username, newEmail } = req.body;
 
   try {
     const updates = {};
 
     if (username) updates.username = username;
-    if (email) updates.email = email;
+    if (newEmail) {
+      const existingUser = await User.findOne({ email: newEmail });
+
+      if (existingUser && existingUser._id.toString() !== req.user.id) {
+        return res.status(409).json({
+          success: false,
+          message: "Email is already in use",
+        });
+      }
+
+      updates.email = newEmail;
+    }
 
     const CurrentUser = await User.findById(req.user.id);
-    const comaprePassword = await bcrypt.compare(oldPassword,CurrentUser.password);
+    const comaprePassword = await bcrypt.compare(currentPassword,CurrentUser.password);
 
     if(!comaprePassword){
       return res.status(400).json({
@@ -90,7 +101,28 @@ export const updateProfile = async (req, res) => {
 };
 
 export const deleteProfile = async (req, res) => {
+  const { username, currentPassword } = req.body;
   try {
+    if(!username || !currentPassword){
+      return res.status(400).json({
+        success: false,
+        message: "Username and password are required",
+      })
+    }
+    const existingUser = await User.findById(req.user.id);
+    const comparePassword = await bcrypt.compare(currentPassword, existingUser.password);
+    if (!comparePassword) {
+      return res.status(400).json({
+        success: false,
+        message: "Incorrect password",
+      });
+    }
+    if(username !== existingUser.username){
+      return res.status(400).json({
+        success: false,
+        message: "Username did not match",
+      })
+    }
     const user = await User.findByIdAndDelete(req.user.id);
     if (!user) {
       return res.status(404).json({
